requirements.txt
google-play-scraper>=1.2.0
PyYAML>=6.0
langdetect>=1.0.9
groq>=0.4.0
python-dotenv>=1.0.0
google-genai>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
APScheduler>=3.10.0
